Use ConvertTo-Json in Connect-vRAServer
Creating JSON using embedded variables in a string can easily break if the user's input contains special characters. In my case, my password would break the encoding and the server would error out. This can be the case, for example, if the password contains a double quote, ends with a backslash, etc...