Allow username to be an email address
We login to vRA using our email addresses. The domain in the email address is not the same as the domain we need to use to login.
By looking at 2x @ in the username, we can assume the first part is an email address.
This way, 'email@address.com@domain.local' could be used to login.